// models/product.ts
import { IPagination } from './pagination';

export interface IProducts extends IPagination {
  products: IProduct[];
}

export interface IProduct {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  brand: string;
  category: string;
  thumbnail: string;
  images: string[];
}